refactor: Firebase 회원가입 로직 signUpUser 서비스 함수로 분리
