+++ src/hooks/auth/useLogin.js
@@ -3,8 +3,7 @@ import { useDispatch } from "react-redux";
 import { useNavigate } from "react-router-dom";
 import { setUser } from "store/userSlice";
 
-import { signIn } from "services/auth";
-import { signInWithGoogle } from "services/auth";
+import { signIn, signInWithGoogle } from "services/auth/auth";
 
 export function useLogin() {
   const [email, setEmail] = useState("");
