+++ src/hooks/auth/useUser.js
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { observeAuth } from "services/auth";
+import { observeAuth } from "services/auth/auth";
 
 export const useUser = () => {
   const [user, setUser] = useState(null);
